fix(ci): split ha regression tests for source budget
